pyproject.toml
[tool.poetry]
name = "articulated"
version = "0.1.0"
description = "Learning Neural Manifold Representations for Articulated Robotics"
authors = [
    "Mathilde Papillon <mpapillon@ucsb.edu>",
    "Francisco Acosta <facosta@ucsb.edu>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# Core ML
torch = ">=2.0.0"
lightning = ">=2.0.0"

# Data & numerics
numpy = ">=1.24.0"
scipy = ">=1.11.0"
h5py = ">=3.9.0"
pyyaml = ">=6.0"

# Visualization
matplotlib = ">=3.7.0"
plotly = ">=5.15.0"

# Dimensionality reduction & analysis
scikit-learn = ">=1.3.0"
umap-learn = ">=0.5.0"

# Reinforcement learning (Stage 3)
gymnasium = {version = ">=0.29.0", extras = ["mujoco"]}
stable-baselines3 = ">=2.0.0"

# Experiment tracking
wandb = ">=0.15.0"

# Notebooks
jupyterlab = ">=4.0.0"
ipywidgets = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
ruff = ">=0.4.0"
mypy = ">=1.10.0"
black = ">=24.0.0"
types-PyYAML = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "B", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
